use concerns to DRY AssociatedUsersControllers
